Organize all router related files under directory
