+++ src/main/__test__/static-file-route.test.ts
@@ -3,7 +3,7 @@
  * Licensed under MIT License. See LICENSE in root directory for more information.
  */
 
-import type { LensApiRequest, Route } from "../router";
+import type { LensApiRequest, Route } from "../router/router";
 import staticFileRouteInjectable from "../routes/static-file-route.injectable";
 import { getDiForUnitTesting } from "../getDiForUnitTesting";
