+++ src/main/routes/metrics/add-metrics-route.injectable.ts
@@ -5,7 +5,7 @@
 
 import { getInjectable } from "@ogre-tools/injectable";
 import { apiPrefix } from "../../../common/vars";
-import type { LensApiRequest, Route } from "../../router";
+import type { LensApiRequest, Route } from "../../router/router";
 import { routeInjectionToken } from "../../router/router.injectable";
 import { ClusterMetadataKey, ClusterPrometheusMetadata } from "../../../common/cluster-types";
 import logger from "../../logger";
