+++ src/main/routes/metrics/get-metric-providers-route.injectable.ts
@@ -5,7 +5,7 @@
 
 import { getInjectable } from "@ogre-tools/injectable";
 import { apiPrefix } from "../../../common/vars";
-import type { Route } from "../../router";
+import type { Route } from "../../router/router";
 import { routeInjectionToken } from "../../router/router.injectable";
 import { PrometheusProviderRegistry } from "../../prometheus";
 import type { MetricProviderInfo } from "../../../common/k8s-api/endpoints/metrics.api";
