+++ src/main/routes/port-forward/get-current-port-forward-route.injectable.ts
@@ -4,7 +4,7 @@
  */
 import { getInjectable } from "@ogre-tools/injectable";
 import { routeInjectionToken } from "../../router/router.injectable";
-import type { LensApiRequest, Route } from "../../router";
+import type { LensApiRequest, Route } from "../../router/router";
 import { apiPrefix } from "../../../common/vars";
 import { PortForward } from "./functionality/port-forward";
